search-example/pyagents/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
async-timeout==5.0.1
bitarray==3.7.1
certifi==2025.8.3
charset-normalizer==3.4.3
coloredlogs==15.0.1
Django==5.2.6
django-scheduler==0.10.1
exceptiongroup==1.3.0
fastembed==0.7.3
filelock==3.19.1
flatbuffers==25.2.10
fsspec==2025.9.0
grpcio==1.74.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.9
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
hyperframe==6.1.0
icalendar==6.3.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
loguru==0.7.3
MarkupSafe==3.0.2
mmh3==5.2.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
onnxruntime==1.22.1
packaging==25.0
pdfreader==0.1.15
pillow==11.3.0
portalocker==3.2.0
protobuf==6.32.0
py_rust_stemmers==0.1.5
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pypdf==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.15.1
redis==6.4.0
regex==2025.9.1
requests==2.32.5
safetensors==0.6.2
schedule==1.2.2
scikit-learn==1.7.1
scipy==1.15.3
sentence-transformers==5.1.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
transformers==4.56.0
triton==3.4.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
